Create a more generic Rules context trait also for context mapping.
